requirements.txt
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
plotly==5.17.0
dash==2.13.0
folium==0.14.0
numpy==1.24.3
pandas==2.0.3
gunicorn==21.2.0
flask-wtf==1.1.1